Move SearchInput to common folder
This component can be reused in other searches (timetables), so it
should rather be in common folder.
