+++ ui/src/components/common/search/index.ts
@@ -1,0 +1,1 @@
+export * from './SearchInput';
